updated pydelorean setup
