+++ setup.py
@@ -5,7 +5,7 @@ from pathlib import Path
 this_directory = Path(__file__).parent
 LONG_DESCRIPTION = (this_directory / "README.rst").read_text()
 
-VERSION = '0.3.1'
+VERSION = '1.0.0'
 DESCRIPTION = 'A package to convert between markup language documents and a forest data structure for efficient processing.'
 
 setup(
@@ -17,9 +17,9 @@ setup(
     long_description=LONG_DESCRIPTION,
     license = "Apache 2.0",
     url = "http://github.com/kj3moraes/delorean",
-    packages = ['pydelorean', 'pydelorean.tree', 'pydelorean.parser'],
+    packages = ['pydelorean', 'pydelorean.tree', 'pydelorean.parser', 'pydelorean.tools'],
     tests_require = ['unittest'],
-    install_requires = ['markdown', 'treelib', 'beautifulsoup4'],
+    install_requires = ['markdown', 'bigtree', 'beautifulsoup4'],
     download_url = 'https://github.com/kj3moraes/delorean/archive/%s.zip' % VERSION,
     classifiers = [
         "Topic :: Utilities",
